+++ apps/frontend/src/libs/components/select/libs/constants/constants.ts
@@ -1,1 +1,2 @@
-export { selectStylesConfig } from "./select-styles-config.constant.js";
+export { SELECT_ICON_LEFT_STYLES_CONFIG } from "./select-icon-left-styles-config.constant.js";
+export { SELECT_STYLES_CONFIG } from "./select-styles-config.constant.js";