+++ apps/frontend/src/libs/modules/map-client/libs/constants/layer-styles.constant.ts
@@ -1,0 +1,12 @@
+const MAP_LAYER_STYLES = {
+	ROUTE: {
+		LINE_COLOR: "hsl(129, 71%, 63%)",
+		LINE_WIDTH: 6,
+	},
+	ROUTE_OUTLINE: {
+		LINE_COLOR: "hsl(162, 91%, 25%)",
+		LINE_WIDTH: 12,
+	},
+};
+
+export { MAP_LAYER_STYLES };