+++ apps/frontend/src/pages/route-details/route-details.tsx
@@ -40,9 +40,10 @@ const RouteDetails = (): React.JSX.Element => {
 	const dataStatus = useAppSelector(
 		({ routeDetails }) => routeDetails.dataStatus,
 	);
+
 	const reviews = useAppSelector(({ routeDetails }) => routeDetails.reviews);
 	const isAuthenticatedUser = Boolean(user);
-	const { control, errors, getValues, handleValueSet } =
+	const { control, errors, getValues, handleReset } =
 		useAppForm<RoutePatchRequestDto>({
 			defaultValues: ROUTE_FORM_DEFAULT_VALUES,
 		});
@@ -57,6 +58,22 @@ const RouteDetails = (): React.JSX.Element => {
 	const handleToggleEditMode = useCallback(() => {
 		setIsEditMode((isEditMode) => !isEditMode);
 	}, []);
+
+	const handleResetFormValues = useCallback(() => {
+		if (!route) {
+			return;
+		}
+
+		handleReset({
+			description: route.description,
+			name: route.name,
+		});
+	}, [handleReset, route]);
+
+	const handleCancel = useCallback(() => {
+		handleResetFormValues();
+		setIsEditMode(false);
+	}, [handleResetFormValues]);
 
 	const handlePatchRequest = useCallback(() => {
 		if (route) {
@@ -76,11 +93,8 @@ const RouteDetails = (): React.JSX.Element => {
 	}, [dispatch, routeId]);
 
 	useEffect(() => {
-		if (route) {
-			handleValueSet("name", route.name);
-			handleValueSet("description", route.description);
-		}
-	}, [route, handleValueSet]);
+		handleResetFormValues();
+	}, [handleResetFormValues]);
 
 	useEffect(() => {
 		void dispatch(routeActions.getReviews({ routeId: Number(routeId) }));
@@ -94,7 +108,11 @@ const RouteDetails = (): React.JSX.Element => {
 	);
 
 	if (dataStatus === DataStatus.PENDING || dataStatus === DataStatus.IDLE) {
-		return <Loader />;
+		return (
+			<div className={styles["loader-container"]}>
+				<Loader />
+			</div>
+		);
 	}
 
 	if (dataStatus === DataStatus.REJECTED) {
@@ -123,7 +141,7 @@ const RouteDetails = (): React.JSX.Element => {
 							/>
 							<div className={styles["edit-mode-controls"]}>
 								<Button label="Save" onClick={handlePatchRequest} />
-								<Button label="Cancel" onClick={handleToggleEditMode} />
+								<Button label="Cancel" onClick={handleCancel} />
 							</div>
 						</>
 					) : (