+++ packages/shared/src/index.ts
@@ -71,6 +71,7 @@ export {
 	type PermissionItemDto,
 	PermissionKey,
 } from "./modules/permission/permission.js";
+export { type PlannedPathResponseDto } from "./modules/planned-paths/planned-paths.js";
 export {
 	type PointOfInterestPatchRequestDto,
 	pointOfInterestUpdateValidationSchema,
@@ -131,6 +132,8 @@ export {
 	type UserAuthResponseDto,
 	UserExceptionMessage,
 	type UserGetByIdItemResponseDto,
+	userGetProfileParametersValidationSchema,
+	type UserPublicProfileResponseDto,
 	UsersApiPath,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,